feat: SSR 적용
