+++ frontend/techpick/src/app/(signed)/layout.tsx
@@ -1,23 +1,49 @@
+import { getBasicFolders } from '@/apis/folder/getBasicFolders';
+import { getFolders } from '@/apis/folder/getFolders';
+import { getTagList } from '@/apis/tag/getTagList';
 import { FeedbackToolbar } from '@/components/FeedbackToolbar';
 import { FolderAndPickDndContextProvider } from '@/components/FolderAndPickDndContextProvider';
 import { ScreenLogger } from '@/components/ScreenLogger';
 import ShortcutKey from '@/components/ShortcutKey';
 import { SideNavigationBar } from '@/components/SideNavigationBar/SideNavigationBar';
+import { getQueryClient } from '@/libs/@react-query/getQueryClient';
+import { folderKeys } from '@/queries/folderKeys';
+import { tagKeys } from '@/queries/tagKeys';
+import { HydrationBoundary, dehydrate } from '@tanstack/react-query';
 import type { PropsWithChildren } from 'react';
 import { pageContainerLayout } from './layout.css';
 
-export default function SignedLayout({ children }: PropsWithChildren) {
+export default async function SignedLayout({ children }: PropsWithChildren) {
+  const queryClient = getQueryClient();
+
+  await Promise.all([
+    queryClient.prefetchQuery({
+      queryKey: folderKeys.basic(),
+      queryFn: getBasicFolders,
+    }),
+    queryClient.prefetchQuery({
+      queryKey: folderKeys.root(),
+      queryFn: getFolders,
+    }),
+    queryClient.prefetchQuery({
+      queryKey: tagKeys.all,
+      queryFn: getTagList,
+    }),
+  ]);
+
   return (
-    <ScreenLogger eventName="page_view_signed_user">
-      <div className={pageContainerLayout}>
-        <FolderAndPickDndContextProvider>
-          {/** 선택한 폴더에 따른 컨텐츠가 나옵니다. */}
-          <SideNavigationBar />
-          {children}
-          <FeedbackToolbar />
-          <ShortcutKey />
-        </FolderAndPickDndContextProvider>
-      </div>
-    </ScreenLogger>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <ScreenLogger eventName="page_view_signed_user">
+        <div className={pageContainerLayout}>
+          <FolderAndPickDndContextProvider>
+            {/** 선택한 폴더에 따른 컨텐츠가 나옵니다. */}
+            <SideNavigationBar />
+            {children}
+            <FeedbackToolbar />
+            <ShortcutKey />
+          </FolderAndPickDndContextProvider>
+        </div>
+      </ScreenLogger>
+    </HydrationBoundary>
   );
 }
